Feat: Added Dummy Customer, Rider, Manger and Admin login for public to view the dashboard of how it works

+++ CreaveIt/client/src/components/adminDashboard/Customers.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-
 import toast from "react-hot-toast";
 import api from "../../config/Api";
 import ManagementTable from "./ManagementTable";
+import { useAuth } from "../../context/AuthContext";
 
 const Customers = () => {
+  const { user } = useAuth();
   const navigate = useNavigate();
-
   const [customers, setCustomers] = useState([]);
   const [loading, setLoading] = useState(false);
   const [deletingId, setDeletingId] = useState(null);
@@ -17,9 +17,7 @@ const Customers = () => {
 
     try {
       const res = await api.get("/admin/customers");
-
       setCustomers(res.data.data || []);
-
       if (showToast) {
         toast.success("Customers refreshed successfully");
       }
@@ -41,21 +39,26 @@ const Customers = () => {
   };
 
   const handleCreate = () => {
+    if (user.email === "admin@gmail.com") {
+      return toast.error("Dummy Admin cannot perfom this action.");
+    }
+
     navigate("/admin-dashboard/add-user");
   };
 
   const handleDelete = async (customer) => {
-    const confirmDelete = window.confirm(`Delete ${customer.fullName}?`);
+    if (user.email === "admin@gmail.com") {
+      return toast.error("Dummy Admin cannot perfom this action.");
+    }
 
+    const confirmDelete = window.confirm(`Delete ${customer.fullName}?`);
     if (!confirmDelete) return;
 
     setDeletingId(customer._id);
 
     try {
       await api.delete(`/admin/customer/${customer._id}`);
-
       setCustomers((prev) => prev.filter((item) => item._id !== customer._id));
-
       toast.success("Customer deleted successfully");
     } catch (error) {
       toast.error(
